Calculate unrotated pixel and reference unrotated mask.
Previously, to use a mask with a rotated image I also padded and rotated
the mask. This took a significant amount of time and space. With this
change the pixel being looked at in the sorting algorithm finds it
related pixel in the mask by unrotating around the image center nd
subtracting the amount of padding added to the x and y in the rotated
image. This improves performance and reuces complexity.