Add arrows indicating sorting angle and direction.
By drawing individual lines to an FBO of the maximum size the window can
be, an image of arrows is generated at runtime and drawn when angle
slider is modified at the specified rotation angle. Resolution of the
arrows can be increased and clarity improved but was having trouble
thickening the line width.